Add initial support for zero-based-indexing.
Enable with the define LUAJIT_ZERO_BASED_INDEXING. When built with
this flag, the following changes are made to the semantics of Lua:

  - The length operator finds an index of a nil preceded by a non-nil.
  - Numeric for loops over a half open interval [start, limit).
  - Table constructors use 0 as the first array key.
  - The ipairs() iterator function returns (1, t[0]), (2, t[1])...

Only the x64 interpreter has been modified so far. Lua code in the
jit/ subfolder has not yet been updated. Library functions other
than ipairs() have not yet been modified.